+++ flextype/Flextype.php
@@ -40,7 +40,7 @@ class Flextype
      *
      * @var string
      */
-    const VERSION = '0.4.2';
+    const VERSION = '0.4.3';
 
     /**
      * Constructor.