+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Flextype 0.4.3, 2018-05-28
+* Content: set text/html request headers for displayCurrentPage() method
+* Content: processCurrentPage() method added
+* Content: event names changed: onPageBeforeRender to onCurrentPageBeforeProcessed
+* Content: event names changed: onPageAfterRender to onCurrentPageAfterProcessed
+* robots.txt file was removed, use Robots plugin instead
+* Code cleanup and refactoring #5
+
 # Flextype 0.4.2, 2018-05-22
 * Settings: cache.enabled is true from now
 * Content: new methods added: initShortcodes() initMarkdown() markdown()